complete show list tutor, sort tutor with favourite and rate
