+++ project_source_code_20120597/lib/features/list_teacher/screens/screen_list_teacher.dart
@@ -2,9 +2,9 @@ import 'package:flutter/material.dart';
 import 'package:lettutor_mobile_toji/commons/components/appbar.dart';
 import 'package:lettutor_mobile_toji/commons/components/sidebar.dart';
 import 'package:lettutor_mobile_toji/commons/const_var.dart';
-import 'package:country_flags/country_flags.dart';
+import 'package:lettutor_mobile_toji/commons/models/tutor_model.dart';
 import 'package:lettutor_mobile_toji/features/list_teacher/components/card_english_type.dart';
-import 'package:lettutor_mobile_toji/features/tutor_profile/screens/tutor_profile.dart';
+import 'package:lettutor_mobile_toji/features/list_teacher/components/item_info_tutor.dart';
 
 class TutorScreen extends StatefulWidget {
   const TutorScreen({Key? key}) : super(key: key);
@@ -31,6 +31,24 @@ class TutorScreenState extends State<TutorScreen> {
     'TOELF',
     'TOEIC'
   ];
+
+  final List<Tutor> listTutor = [
+    Tutor(id: '1', name: 'Toai', avatar: 'assets/common/img_user.png', introduce: 'introduce', rate: 5, numberRate: 1500, countryCode: 'US', education: 'education', language: ['English'], specialized: ['specialized'], videoPath: 'videoPath', interest: 'interest', teachingExperience: 'teachingExperience', isFavourite: false),
+    Tutor(id: '2', name: 'Toai', avatar: 'assets/common/img_user.png', introduce: 'introduce', rate: 3.7, numberRate: 1500, countryCode: 'US', education: 'education', language: ['English'], specialized: ['specialized'], videoPath: 'videoPath', interest: 'interest', teachingExperience: 'teachingExperience', isFavourite: false),
+  ];
+
+  @override
+  void initState() {
+    super.initState();
+    listTutor.sort((Tutor a, Tutor b) {
+      // compare isFavourite and rate
+      if (a.isFavourite == b.isFavourite) {
+        return b.rate.compareTo(a.rate);
+      } else {
+        return a.isFavourite! ? -1 : 1;
+      }
+    });
+  }
 
   void _openDrawer() {
     _scaffoldKey.currentState?.openDrawer();
@@ -163,218 +181,16 @@ class TutorScreenState extends State<TutorScreen> {
                           style: TextStyle(
                               fontWeight: FontWeight.bold, fontSize: 25),
                         ),
-                        const Padding(
-                          padding: EdgeInsets.all(1),
-                          child: GeneralInfoTutor(),
+                        Padding(
+                          padding: const EdgeInsets.all(1),
+                          child: Column(
+                            children: listTutor.map((e) => GeneralInfoTutor(tutor: e)).toList(),
+                          ),
                         )
                       ],
                     )),
               )
             ],
-          ),
-        ));
-  }
-}
-
-class GeneralInfoTutor extends StatefulWidget {
-  const GeneralInfoTutor({super.key});
-
-  @override
-  State<StatefulWidget> createState() {
-    // ignore: no_logic_in_create_state
-    return GeneralInfoTutorState(
-        avatar: Image.asset('assets/common/img_user.png'),
-        name: 'Not Found',
-        rate: 3,
-        cardEnglishTypeData: [
-          'EnglishToji',
-          'English2',
-          'English3',
-          'English1',
-          'English2',
-          'English3'
-        ]);
-  }
-}
-
-class GeneralInfoTutorState extends State<GeneralInfoTutor> {
-  final Image avatar;
-  final String name;
-  final int rate;
-  final List<String> cardEnglishTypeData;
-
-  GeneralInfoTutorState(
-      {required this.avatar,
-      required this.name,
-      required this.rate,
-      required this.cardEnglishTypeData});
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-        padding: const EdgeInsets.all(5),
-        child: Container(
-          decoration: BoxDecoration(
-            borderRadius: BorderRadius.circular(20),
-            boxShadow: const [
-              BoxShadow(
-                color: Colors.black,
-                blurRadius: 2.0,
-              )
-            ],
-          ),
-          child: Container(
-            decoration: BoxDecoration(
-              borderRadius: BorderRadius.circular(20),
-              color: Colors.white,
-            ),
-            child: Column(
-              children: <Widget>[
-                Padding(
-                  padding: const EdgeInsets.fromLTRB(0, 0, 0, 0),
-                  child: Row(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: <Widget>[
-                      Expanded(flex: 1, child: Container()),
-                      Expanded(
-                          flex: 6,
-                          child: Column(
-                            children: <Widget>[
-                              Padding(
-                                padding: const EdgeInsets.fromLTRB(0, 10, 0, 0),
-                                child: SizedBox(
-                                  height: 80,
-                                  width: 80,
-                                  child: avatar,
-                                ),
-                              ),
-                              Padding(
-                                padding: const EdgeInsets.all(1),
-                                child: Text(
-                                  name,
-                                  style: const TextStyle(
-                                      fontWeight: FontWeight.bold,
-                                      fontSize: 16),
-                                ),
-                              ),
-                              Padding(
-                                padding: const EdgeInsets.all(1),
-                                child: Row(
-                                  mainAxisAlignment: MainAxisAlignment.center,
-                                  children: <Widget>[
-                                    SizedBox(
-                                      width: 30,
-                                      height: 30,
-                                      child: CountryFlag.fromCountryCode('ES'),
-                                    ),
-                                    const Padding(
-                                        padding:
-                                            EdgeInsets.fromLTRB(10, 0, 0, 0),
-                                        child: Text('ES'))
-                                  ],
-                                ),
-                              ),
-                              Padding(
-                                  padding: const EdgeInsets.all(1),
-                                  child: Row(
-                                      mainAxisAlignment:
-                                          MainAxisAlignment.center,
-                                      children: List.generate(5, (index) {
-                                        return GestureDetector(
-                                          child: Icon(
-                                            Icons.star,
-                                            size: 30,
-                                            color: index < rate
-                                                ? Colors.amber
-                                                : const Color.fromARGB(
-                                                    255, 78, 78, 78),
-                                          ),
-                                        );
-                                      }))),
-                              Padding(
-                                padding: const EdgeInsets.all(1),
-                                child: Wrap(
-                                  alignment: WrapAlignment.center,
-                                  spacing: 5,
-                                  runSpacing: 5,
-                                  children: cardEnglishTypeData
-                                      .map((e) => CardEnglishType(
-                                            englishType: e,
-                                            onPressed: () {},
-                                          ))
-                                      .toList(),
-                                ),
-                              ),
-                              const Padding(
-                                  padding: EdgeInsets.all(1),
-                                  child: Text(
-                                    'I am passionate about running and fitness, I often compete in trail/mountain running events and I love pushing myself. I am training to one day take part in ultra-endurance events.',
-                                    textAlign: TextAlign.center,
-                                  )),
-                              Padding(
-                                  padding:
-                                      const EdgeInsets.fromLTRB(0, 5, 0, 10),
-                                  child: Container(
-                                    width: 120,
-                                    height: 45,
-                                    decoration: BoxDecoration(
-                                      border: Border.all(
-                                        color:
-                                            primaryColor, // Color of the border
-                                        width: 1.5, // Width of the border
-                                      ),
-                                      borderRadius: BorderRadius.circular(
-                                          16), // Border radius for rounded corners
-                                    ),
-                                    child: TextButton(
-                                      onPressed: () {
-                                        Navigator.push(
-                                          context,
-                                          MaterialPageRoute(
-                                            builder: (context) =>
-                                                TutorProfile(),
-                                          ),
-                                        );
-                                      },
-                                      child: const Row(
-                                        mainAxisAlignment:
-                                            MainAxisAlignment.center,
-                                        children: <Widget>[
-                                          Icon(
-                                            Icons.calendar_month,
-                                            size: 25,
-                                            color: primaryColor,
-                                          ),
-                                          Padding(
-                                            padding:
-                                                EdgeInsets.fromLTRB(8, 0, 0, 0),
-                                            child: Text(
-                                              'Đặt lịch',
-                                              style: TextStyle(
-                                                  color: primaryColor,
-                                                  fontSize: 16),
-                                            ),
-                                          )
-                                        ],
-                                      ),
-                                    ),
-                                  ))
-                            ],
-                          )), // Avatar
-                      const Expanded(
-                          flex: 1,
-                          child: Padding(
-                              padding: EdgeInsets.fromLTRB(0, 10, 10, 0),
-                              child: Icon(
-                                Icons.favorite_border,
-                                color: primaryColor,
-                              ))), // Heart
-                    ],
-                  ),
-                ),
-                const Padding(padding: EdgeInsets.fromLTRB(0, 0, 0, 0))
-              ],
-            ),
           ),
         ));
   }
