+++ project_source_code_20120597/lib/features/tutor_profile/screens/tutor_profile.dart
@@ -20,7 +20,7 @@ class TutorProfileState extends State<TutorProfile> {
       avatar: 'assets/common/img_user.png',
       introduce:
           'I am passionate about running and fitness, I often compete in trail/mountain running events and I love pushing myself. I am training to one day take part in ultra-endurance events. I also enjoy watching rugby on the weekends, reading and watching podcasts on Youtube. My most memorable life experience would be living in and traveling around Southeast Asia.',
-      rate: 1,
+      rate: 4.7,
       numberRate: 2,
       countryCode: 'US',
       education: 'BA',
